Rename telemetry to tracing
Telemetry contains both metrics and tracing, and we make a clear distinction between them in our modules.
This will allow us to add a telemetry submodule that contains code used to setup the OpenTelemetry library.
