+++ language-support/scala/bindings-akka/BUILD.bazel
@@ -57,7 +57,7 @@ da_scala_library(
         "//ledger/ledger-api-client",
         "//ledger/ledger-api-common",
         "//ledger/ledger-api-domain",
-        "//observability/telemetry",
+        "//observability/tracing",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:ch_qos_logback_logback_core",
         "@maven//:com_google_api_grpc_proto_google_common_protos",
@@ -99,7 +99,7 @@ da_scala_test_suite(
         "//ledger-api/rs-grpc-bridge",
         "//ledger/ledger-api-client",
         "//ledger/metrics",
-        "//observability/telemetry",
+        "//observability/tracing",
         "@maven//:com_google_api_grpc_proto_google_common_protos",
         "@maven//:com_typesafe_config",
         "@maven//:org_scalatest_scalatest_compatible",
