+++ ledger/ledger-api-client/BUILD.bazel
@@ -34,7 +34,7 @@ da_scala_library(
         "//libs-scala/ports",
         "//libs-scala/resources",
         "//observability/metrics",
-        "//observability/telemetry",
+        "//observability/tracing",
         "@maven//:com_google_api_grpc_proto_google_common_protos",
         "@maven//:com_google_protobuf_protobuf_java",
         "@maven//:io_dropwizard_metrics_metrics_core",
@@ -72,7 +72,7 @@ da_scala_test_suite(
         "//libs-scala/contextualized-logging",
         "//libs-scala/grpc-utils",
         "//observability/metrics",
-        "//observability/telemetry",
+        "//observability/tracing",
         "@maven//:com_google_api_grpc_proto_google_common_protos",
         "@maven//:com_google_protobuf_protobuf_java",
         "@maven//:io_dropwizard_metrics_metrics_core",
@@ -123,7 +123,7 @@ da_scala_test_suite(
         "//libs-scala/ports",
         "//libs-scala/resources",
         "//observability/metrics",
-        "//observability/telemetry",
+        "//observability/tracing",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:io_netty_netty_handler",
     ],
