+++ ledger/ledger-api-common/BUILD.bazel
@@ -51,7 +51,7 @@ da_scala_library(
         "//libs-scala/resources-grpc",
         "//libs-scala/timer-utils",
         "//observability/metrics",
-        "//observability/telemetry",
+        "//observability/tracing",
         "@maven//:commons_codec_commons_codec",
         "@maven//:commons_io_commons_io",
         "@maven//:io_grpc_grpc_netty",
@@ -140,8 +140,8 @@ da_scala_test_suite(
         "//libs-scala/logging-entries",
         "//libs-scala/scala-utils",
         "//observability/metrics",
-        "//observability/telemetry",
-        "//observability/telemetry:telemetry-test-lib",
+        "//observability/tracing",
+        "//observability/tracing:tracing-test-lib",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:ch_qos_logback_logback_core",
         "@maven//:com_google_api_grpc_proto_google_common_protos",
