+++ ledger/ledger-api-common/src/test/suite/scala/com/digitalasset/platform/server/api/services/grpc/GrpcCommandSubmissionServiceSpec.scala
@@ -14,7 +14,7 @@ import com.daml.lf.data.Ref
 import com.daml.logging.LoggingContext
 import com.daml.metrics.Metrics
 import com.daml.platform.server.api.services.domain.CommandSubmissionService
-import com.daml.telemetry.{SpanAttribute, TelemetryContext, TelemetrySpecBase}
+import com.daml.tracing.{SpanAttribute, TelemetryContext, TelemetrySpecBase}
 import org.mockito.captor.ArgCaptor
 import org.mockito.{ArgumentMatchersSugar, MockitoSugar}
 import org.scalatest.matchers.should.Matchers
