+++ ledger/participant-integration-api/BUILD.bazel
@@ -73,7 +73,7 @@ compile_deps = [
     "//libs-scala/timer-utils",
     "//libs-scala/nameof",
     "//libs-scala/struct-json/struct-spray-json",
-    "//observability/telemetry",
+    "//observability/tracing",
     "//observability/metrics",
     "@maven//:com_google_guava_guava",
     "@maven//:com_zaxxer_HikariCP",
@@ -320,8 +320,8 @@ da_scala_test_suite(
         "//libs-scala/timer-utils",
         "//observability/metrics",
         "//observability/metrics:metrics-test-lib",
-        "//observability/telemetry",
-        "//observability/telemetry:telemetry-test-lib",
+        "//observability/tracing",
+        "//observability/tracing:tracing-test-lib",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:ch_qos_logback_logback_core",
         "@maven//:com_github_ben_manes_caffeine_caffeine",
