+++ ledger/participant-state/src/main/scala/com/daml/ledger/participant/state/v2/WriteConfigService.scala
@@ -9,7 +9,7 @@ import com.daml.ledger.configuration.Configuration
 import com.daml.lf.data.Ref
 import com.daml.lf.data.Time.Timestamp
 import com.daml.logging.LoggingContext
-import com.daml.telemetry.TelemetryContext
+import com.daml.tracing.TelemetryContext
 
 trait WriteConfigService {
