+++ ledger/recovering-indexer-integration-tests/BUILD.bazel
@@ -55,7 +55,7 @@ da_scala_test_suite(
         "//libs-scala/resources-grpc",
         "//libs-scala/timer-utils",
         "//observability/metrics",
-        "//observability/telemetry",
+        "//observability/tracing",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:io_dropwizard_metrics_metrics_core",
         "@maven//:org_mockito_mockito_core",
