+++ ledger/sandbox-on-x/BUILD.bazel
@@ -61,7 +61,7 @@ da_scala_library(
         "//libs-scala/resources-akka",
         "//libs-scala/resources-grpc",
         "//observability/metrics",
-        "//observability/telemetry",
+        "//observability/tracing",
         "@maven//:com_google_guava_guava",
         "@maven//:com_google_protobuf_protobuf_java",
         "@maven//:com_typesafe_config",
