+++ observability/tracing/BUILD.bazel
@@ -8,12 +8,12 @@ load(
 )
 
 da_scala_library(
-    name = "telemetry",
+    name = "tracing",
     srcs = glob(["src/main/scala/**/*.scala"]),
     resources = glob(["src/main/resources/**/*"]),
     scala_deps = [
     ],
-    tags = ["maven_coordinates=com.daml:telemetry:__VERSION__"],
+    tags = ["maven_coordinates=com.daml:tracing:__VERSION__"],
     visibility = [
         "//visibility:public",
     ],
@@ -25,19 +25,19 @@ da_scala_library(
 )
 
 da_scala_library(
-    name = "telemetry-test-lib",
+    name = "tracing-test-lib",
     srcs = glob(["src/test/lib/scala/**/*.scala"]),
     scala_deps = [
         "@maven//:org_scalactic_scalactic",
         "@maven//:org_scalatest_scalatest_core",
     ],
-    tags = ["maven_coordinates=com.daml:telemetry-test-lib:__VERSION__"],
+    tags = ["maven_coordinates=com.daml:tracing-test-lib:__VERSION__"],
     visibility = [
         "//visibility:public",
     ],
     runtime_deps = [],
     deps = [
-        ":telemetry",
+        ":tracing",
         "@maven//:io_opentelemetry_opentelemetry_api",
         "@maven//:io_opentelemetry_opentelemetry_context",
         "@maven//:io_opentelemetry_opentelemetry_sdk_testing",
@@ -47,7 +47,7 @@ da_scala_library(
 )
 
 da_scala_test_suite(
-    name = "telemetry-tests",
+    name = "tracing-tests",
     size = "small",
     srcs = glob(["src/test/suite/scala/**/*.scala"]),
     scala_deps = [
@@ -59,8 +59,8 @@ da_scala_test_suite(
         "@maven//:org_scalatest_scalatest_wordspec",
     ],
     deps = [
-        ":telemetry",
-        ":telemetry-test-lib",
+        ":tracing",
+        ":tracing-test-lib",
         "//ledger-api/rs-grpc-akka",
         "//ledger-api/rs-grpc-akka:rs-grpc-akka-tests-lib",
         "//ledger-api/rs-grpc-bridge",
