+++ observability/tracing/src/main/scala/com/daml/tracing/SpanAttribute.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2023 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.telemetry
+package com.daml.tracing
 
 import io.opentelemetry.api.common.AttributeKey
