+++ observability/tracing/src/main/scala/com/daml/tracing/SpanName.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2023 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.telemetry
+package com.daml.tracing
 
 object SpanName {
   val LedgerConfigProviderInitialConfig = "daml.ledger.config-provider.initial-config"
