+++ observability/tracing/src/main/scala/com/daml/tracing/package.scala
@@ -6,6 +6,6 @@ package com.daml
 import io.opentelemetry.api.GlobalOpenTelemetry
 import io.opentelemetry.api.trace.Tracer
 
-package object telemetry {
+package object tracing {
   val OpenTelemetryTracer: Tracer = GlobalOpenTelemetry.getTracer("com.daml.telemetry")
 }
